requirements.txt
aiohttp
ujson
aiogram
youtube-dl
Pillow
